feat: Implement LandRoverState to launch rovers to Plateau
Implemented a new state to launch rovers to Plateau. This state shows
a map of the current Plateau and the rovers that are currently on it.
The user can either add a new rover through a series of prompts,
or exit the program from this state. It is highly experimental and
subject to change.